requirements.txt
absl-py==2.3.1
altgraph==0.17.5
astunparse==1.6.3
beautifulsoup4==4.14.3
blinker==1.9.0
bottle==0.13.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
clr_loader==0.2.9
colorama==0.4.6
contourpy==1.3.2
curl_cffi==0.13.0
cycler==0.12.1
dnspython==2.8.0
email-validator==2.3.0
Flask==3.1.2
flask-cors==6.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
flatbuffers==25.9.23
fonttools==4.61.1
frozendict==2.4.7
gast==0.7.0
google-pasta==0.2.0
greenlet==3.3.1
grpcio==1.76.0
gunicorn==24.1.1
h5py==3.15.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
keras==3.12.0
kiwisolver==1.4.9
libclang==18.1.1
lightgbm==4.6.0
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
multitasking==0.0.12
namex==0.1.0
numpy==1.26.4
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pandas==2.3.3
peewee==3.18.3
pefile==2024.8.26
pillow==12.0.0
platformdirs==4.5.1
protobuf==6.33.2
proxy_tools==0.1.0
psycopg2-binary==2.9.11
pycparser==2.23
Pygments==2.19.2
pyinstaller==6.17.0
pyinstaller-hooks-contrib==2025.10
pyparsing==3.3.1
python-dateutil==2.9.0.post0
pythonnet==3.0.5
pytz==2025.2
pywebview==6.1
pywin32-ctypes==0.2.3
requests==2.32.5
rich==14.2.0
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
soupsieve==2.8
SQLAlchemy==2.0.46
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
tensorflow_keras==0.1
termcolor==3.2.0
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.0
websockets==15.0.1
Werkzeug==3.1.4
wrapt==2.0.1
xgboost==3.1.2
yfinance==0.2.66
